+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed error when receiving JSON responses that include `"t"` as a key.
 - Allow underscores (`_`) in HTTP header names.
 - Fixed <kbd>C-u C-c C-r C-r</kbd> and <kbd>C-u C-c C-r C-s</kbd> not displaying the response buffer when a response is received.
+- Fixed Verb not reading parent headings when buffer has been narrowed (e.g. with `org-narrow-to-subtree`).
 
 ## **2.13.1** - 2021-01-12 (MELPA Stable)
 - Heading properties (e.g. `Verb-Store`, `Verb-Map-Request`) can now be inherited from parent headings by setting `org-use-property-inheritance` to `t`.